new fucking with even worse fucking code
