+++ CCCP/include/cccp.h
@@ -4,7 +4,6 @@
 #include "../include/main.h"
 #include "../include/make.h"
 #include "../include/spm.h"
-#include "../include/log.h"
 #include "../include/pkg.h"
 #include "../include/data.h"
