+++ CCCP/lib/log.hpp
